Repair AuthenticationTests#loginSucceeds
loginSucceeds was failing because Thymeleaf rendering was not
happening, which means that the csrf token couldn't be parsed by the
test. It appeared to only be parsed for the benefit of the ensuing
POST, so I just removed it, replacing it with Spring Security's test
support for csrf tokens.

Thereafter, there was a Spring Session serialization error, which was
caused by the Mockito UserDetails object. I replaced this with a
simple POJO, using the User and Consumer domain objects respectively.

The test now passes.