Introduce authorization_code Redirect Test
The existing login test works great to confirm that an individual can
login to the Authorization Server. But, it does not confirm that
the redirection, which happens as part of the OAuth 2.0 Authorization
flow actually works.

We can supply a separate test that hits the /authorize endpoint with a
logged in user to show how the app will behave when that endpoint is
hit.